MainController/requirements.txt
esphome
# snap install mqtt-explorer
# sudo apt install -y mosquitto
paho-mqtt
tornado
esptool
grpcio-tools
#python -m  grpc_tools.protoc -I protos\ --python_out=. --grpc_python_out=. protos\ShadeShell.proto
# sudo chmod a+rw /dev/ttyUSB0
